feat: add safe signature flow to user controller
Without this patch the user controller only allows updating the user
data directly via EOA. Calling it with a multisig like the Safe would
fail as the smart wallet isn't able to create a signature.

In this instance the API needs to track signature requests, so that the
user data can be written once the signature threshold is met.

The need to track these signature requests arises from the fact that you
cannot delete signatures from the Safe once they were "used". They will
stick around forever and it is up to us to track which signature
corresponds to which change, if it is discarded or implemented, etc.

The patch introduces a new folder: src/lib/users. The reason is that the
project is currently structured by implementation details, so I wanted
to keep the user controller in with the rest of the controllers but have
a place to store this business logic.