feat: add signature request queue processing
Without this patch we're missing a way for the API to execute the
signature requests and commit the data to record if the signature
threshold is met.

This patch adds the request POST /signature-requests/process which reads
all **pending** signature requests from the data DB, checks if the
confirmations meet the signature threshold and if so writes the attached
username and avatar changes to the users table.

To stay within the rate limit of 5 requests per second required by the
Safe API I implemented a rate limited queue and worker that process them
concurrently.

The request doesn't require authentication, so it could be invoked by a
CRON job, as well as the user when they log in or specifically request
an update.