feat: run cron job to start signature request processing
Without this patch we rely on users making a POST request to
/v1/signature-requests/process to kick off signature request processing.

Since it's not going to be high load, we want to have a regular cron job
that relieves users of that duty and will run signature request
validations on autopilot.

This way users just have to wait a bit to find their signature requests
validated once the Safe message has reached the required threshold.

Also made a small change to the log statement in the processor to make
it more clear that it's running signature requests, now that a console
log will show up every 30 seconds.