fix(metadata.test.ts): metadata validation handling and responses
Restores the metadata validation endpoint testing. Additionally the full
metadata controller is using try catches and we updated the response
types. The response types have been made more simple and flexible, for
one to support the cases of validation where the processing can be
successful but the data is invalid. Lastly, responses where data is set
as null were cleaned so simply not return data.