refactor: make env var assertion more intuitive
The problem right now is that the constants that are loaded from the
environment have an additional name if specified. This additional name
is very confusing if you're presented with the error message

Environment variable Alchemy API key is not set.

This makes you wonder if perhaps you have mistyped the variable name.
It's almost always better to just print the variable name as it is
expected in your .env file. One notable exception is perhaps our Web3Up
proof and secret.

That's why this patch changes the approach of assertExists() to only
deal with environment variables (it's not used in any other way
currently) and have the caller specify the variable name and an optional
hint to help the person stumbling over a missing env var error with
additional context.

The name of assertExists() was renamed to getRequiredEnvVar() to better
reflect what the function is doing. The filename was also renamed to
further support this.