test: simplify verifyAuthSignedData test
The current test file was testing library code that is already
comprehensively tested. It was also making slow RPC requests and slowed
the test suite down substantially. It requires the Alchemy API key to be
set which is not something that we want, so this patch is doing away
with all superfluous tests and focuses on testing only the code that
we've added around the library call.

Also lowered the coverage thresholds since the EVM client getter is now
mocked and its internals are thus not covered by these tests any longer.