test: mock RPC URL module in Safe sig verifier tests
Without this patch the Safe signature verification code will get RPC
URLs for the specified chain id. This in turn will load the Alchemy API
key, which is not relevant for our tests.

This patch mocks all of this away, so that the API key doesn't need to
be present in the environment while running tests.

Also lowered the coverage threshold of functiosn to 52 as mocking
getRpcUrl() doesn't touch the internals of this function. This should be
thoroughly tested in a dedicated test file anyway.