chore: exclude services/ from coverage
Our services are difficult to test at the moment and we don't really
have tests for them. However, when you make changes to the GraphQL
schema and potentially introduce new entities, you'll also alway have to
add code to the services and might go below the thresholds set.

PR #218 is going to address this problem and make the services as well
as some of the GraphQL testable.

This is in anticipation of the next commit which is going to introduce
the collections entity as a top-level entity.