feat: implement SafeRPCStrategy
For signature validations we need to make contract read calls. Safe
protocol kit expects an RPC URL to make these calls, but on Filecoin
mainnet and Filecoin Calibration they need to be JWT authenticated. This
poses a problem with the initializer of the Safe SDK so we decided to
use public RPCs on these two networks.