feat(supabaseservice): refactor base service to use strategies
Refactors the base supabase service to apply strategies from
`QueryStrategies.ts` when queried. This simplifies the architecture of
the base service to have cleaner code and improve the testability and
typing of the codebase.

This is a setup for introducing more robust tests and removes the
abundance of any-typed fields